host_net: add transport and raw layer ops
The implementation is described by the following diagram:

|---------------------------------------|
|              User                     |
|---------------------------------------|
| (Rel-Go-back-N)   (Bypass)            |   <-   Transport Layer
|---------------------------------------|
| (Raw Socket)  (Raw IB Verbs)  (DPDK)  |  <-   Raw Net Layer
|---------------------------------------|
|         Ethernet NIC                  |
|---------------------------------------|

In this commit, we abstracted two layers: transport and raw_net.

Transport is where we do reliable transmission. For now we have
go-back-N, and bypass.

Raw net is the layer where talk with NIC directly.
For now we have raw socket and raw verbs. DPDK is possible too.